fixing python state to int16

+++ include/hubo.h
@@ -211,9 +211,9 @@ typedef struct hubo_sensor_param {
 	uint16_t boardNo;	///< Sensor Board Nuber
 	uint8_t active;		///< Active sensor
 	char name[5];		///< Name of sensor
-        int8_t xsign;
-        int8_t ysign;
-        int8_t zsign;
+        int16_t xsign;
+        int16_t ysign;
+        int16_t zsign;
 }__attribute__((packed)) hubo_sensor_param_t;
 
 typedef struct hubo_joint_param {
@@ -225,7 +225,7 @@ typedef struct hubo_joint_param {
 	uint16_t harmonic;	///< gear ratio of harmonic drive
 	uint16_t enc;		///< encoder size
 	uint16_t jmc;		///< motor controller number
-	int8_t dir;		    ///< direction
+	int16_t dir;		    ///< direction
 	uint8_t can;		///< can channel
 	uint8_t numMot;		///< number of motors
 	char name[4];		///< name
